Added Earned field for house when sold

+++ app/src/main/java/com/example/auctionhouseapp/Objects/AuctionHouse.kt
@@ -109,7 +109,7 @@ class AuctionHouse: User , Serializable {
         if (TotalRaters == 0)
             Rating = 0.0
         else
-            Rating = (Data[Constants.HOUSE_RATING_SUM] as Long).toDouble()/TotalRaters
+            Rating = (Data[Constants.HOUSE_RATING_SUM] as Double)/TotalRaters
         NextSalesDay = (Data[Constants.HOUSE_NEXT_SALES_DATE] as Timestamp?)?.toDate()
         profile_img_url = Data[Constants.PROFILE_URL] as String?
     }
